# contracts/openapi/pramaan-backend.yaml
openapi: 3.1.0
info:
  title: Pramaan Next — Backend (Phase 1)
  version: 0.1.0
  description: >
    Evidence ingestion, versioning, integrity, deterministic repository queries,
    application metadata, and the collection scheduler. Phase 1 uses deterministic
    mock enterprise integrations. No LLM. Source of truth for backend ⇄ frontend
    and backend ⇄ agents interfaces — implement against this file, do not hand-roll.
servers:
  - url: http://localhost:8080
    description: Local dev
...
paths:
...
  /api/v1/controls/{id}:
    parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
    get:
      tags: [controls]
      summary: Get one control
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/ControlView" } } } }
        "404": { $ref: "#/components/responses/Error" }
...
  /api/v1/applications:
    get:
      tags: [applications]
      summary: List applications (metadata)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ApplicationView" }
    post:
      tags: [applications]
      summary: Create or replace an application
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ApplicationUpsert" }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/ApplicationView" } } } }
  /api/v1/applications/{slug}:
    parameters:
      - name: slug
        in: path
        required: true
        schema: { type: string, pattern: "^[a-z0-9][a-z0-9-]{1,118}$" }
    get:
      tags: [applications]
      summary: Get one application
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/ApplicationView" } } } }
        "404": { $ref: "#/components/responses/Error" }
    put:
      tags: [applications]
      summary: Update an application (slug in path and body must match)
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/ApplicationUpsert" } } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/ApplicationView" } } } }
...
  /api/v1/admin/users:
    get:
      tags: [admin]
      summary: List admin personas (no credentials stored)
      responses:
        "200": { description: OK, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/AdminUserView" } } } } }
    post:
      tags: [admin]
      summary: Create or replace a persona
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/AdminUserUpsert" } } }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/AdminUserView" } } } }
        "400": { $ref: "#/components/responses/Error" }
  /api/v1/admin/users/{username}:
    parameters: [ { name: username, in: path, required: true, schema: { type: string } } ]
    get:
      tags: [admin]
      summary: Get one persona
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/AdminUserView" } } } }
        "404": { $ref: "#/components/responses/Error" }
    put:
      tags: [admin]
      summary: Update a persona (username in path and body must match)
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/AdminUserUpsert" } } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/AdminUserView" } } } }
        "400": { $ref: "#/components/responses/Error" }
    delete:
      tags: [admin]
      summary: Remove a persona
      responses:
        "204": { description: No Content }
        "404": { $ref: "#/components/responses/Error" }
  /api/v1/admin/users/{username}/active:
    parameters: [ { name: username, in: path, required: true, schema: { type: string } } ]
    put:
      tags: [admin]
      summary: Enable or disable a persona
      parameters: [ { name: value, in: query, required: true, schema: { type: boolean } } ]
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/AdminUserView" } } } }
        "404": { $ref: "#/components/responses/Error" }
components:
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorBody" }
  schemas:
    ErrorBody:
      type: object
      properties:
        timestamp: { type: string, format: date-time }
        status: { type: integer }
        error: { type: string }
        message: { type: string }
...
    AdminUserUpsert:
      type: object
      required: [username, displayName, roles]
      properties:
        username: { type: string, pattern: "[a-z0-9][a-z0-9._-]{1,63}" }
        displayName: { type: string }
        email: { type: string, format: email, nullable: true }
        roles: { type: array, items: { type: string } }
        active: { type: boolean, nullable: true }
    AdminUserView:
      type: object
      properties:
        username: { type: string }
        displayName: { type: string }
        email: { type: string, nullable: true }
        roles: { type: array, items: { type: string } }
        active: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    ApplicationUpsert:
      type: object
      required: [slug, name]
      properties:
        slug: { type: string, pattern: "^[a-z0-9][a-z0-9-]{1,118}$" }
        name: { type: string }
        businessUnit: { type: string }
        criticality: { type: string, enum: [LOW, MEDIUM, HIGH, CRITICAL] }
        owner: { type: string }
        technology: { type: array, items: { type: string } }
    ApplicationView:
      allOf:
        - $ref: "#/components/schemas/ApplicationUpsert"
        - type: object
          properties:
            autoCreated: { type: boolean }
            createdAt: { type: string, format: date-time }
            updatedAt: { type: string, format: date-time }
...
    ControlView:
      type: object
      properties:
        id: { type: string }
        framework: { type: string }
        title: { type: string }
        domain: { type: string }
...